Refactor server to use new listener class
This wraps up the socket and its management thread, allowing each
piece to be better encapsulated and specific.
Also, a first draft of the new Message class is in place as stated in
issue #7.  This will allow the server's client connection and
management logic to be moved to an external class, along with
allowing more error-checking and handling in that logic.