Move rewrite along to mostly functioning status
The Client GUI appears to hang when starting the main window.  Not
entirely sure what the problem is yet, but the server components are
tested and appear to have no problems.

The ObservedList class was written hastily and may not be functioning
entirely as intended.  Also, the Client-ClientViewController interface
was adjusted due to the rewrite of the Client.  The error should lie
in one of these two places.