Fix stopping and shutting down bugs.
The shutdown issue was actually in client connection, because I never
shut down the connections from the server side.  It only appeared
when a client had connected, which is what made me look in that
direction.

Closes #2.
Closes #8.
Except for in the case of issue #13 where a reference to a shadowed
client connection is lost and cannot be shut down.