Add MessageProcessor and put it to use
Now instead of different places handling receiving  and decoding a
message header and maybe a message, this is only done by
MessageProcessor.  Also fixed a bug in the decoding of messages from
bytes.  I had improperly assumed struct.unpack was decoding bytes to a
string, but this was not the case.