Finish adding message processing to client/server
Now both the client and server connections receive messages, then
places those messages in a processing queue, where a MessageProcessor
finally calls the appropriate MessageHandler method in response to the
MessageType.  Both Client and Server subclass MessageHandler, allowing
them to override the messages that concern their respective ends of the
communication and throw warnings for the rest.

Closes #12.
Both #1 and #10 are set from a message processing standpoint, but need
GUI enhancements to allow users to trigger the behavior.