Add handling for 2 of 3 username exceptions
The issue with UsernameRejectedException is that it is raised by the
Client without prompt from the GUI.  There is not an obvious point to
wrap up in a try/except block and handle the exception.  Also, this
ends the mainloop of the MessageProcessor for that client, making it
unable to process any incoming messages.