Add a new queue between Client/Server-controllers
Now, instead of using the ObservedKeyListDict and ObservedList(along
with a few set functions in Client) there is a system message queue.
The observer pattern code worked fine when there was a piece of data
to monitor, but on more obscure events like reacting to a shutdown or
a username rejection the tie-ins were less useful.

Now, using the new set of SystemMessage classes modeled after the
Message classes, Client/Server communicate to their unknown
ViewController by placing SystemMessages in this queue to be processed.

The code in Client/Server and their respective ViewControllers needs
to be cleaned up, but this allows a more easily expandable channel for
communication.

Closes #31.