Add dynamic creation of default handle_ methods
Now, [System]MessageHandler have all of the default "raise_warning"
methods created dynamically based on the values in
[System]MessageType.  The conditional logic in
[System]MessageProcessor is also generated dynamically in a dictionary
used for looking up the proper method to call.

The next step is to find a way to properly bind the functions within
the handler class definition, so that I can pass the respective type
to its initializer and no longer need separate classes for system and
non-system messages.