Update Handler/Processor to work with SystemMsgs
MessageHandler and Processor now expect an argument of either
MessageType or SystemMessageType.  Then it uses that to dynamically add
the appropriate default handle methods to the MessageHandler and the
logic for calling the appropriate handle method depending on message
type to MessageProcessor.

Also updates a few more of the classes to use the leading underscore
for "private" method names.